day 15 part one real messy

+++ src/utils/string_utils.h
@@ -8,7 +8,7 @@
 namespace string {
     auto splitString(std::string_view toSplit, char delimiter) -> std::vector<std::string_view>;
 
-    auto toInt(std::string_view input) -> size_t ;
+    auto toInt(std::string_view input) -> long;
 }
 
 #endif // ADVENT_OF_CODE_2022_CPP_STRING_UTILS_H
